comments on proposal
Great job!  I would suggest that your final product be an .Rmd script that Kathy and future students could run in future to update whenever new data come in.  That .Rmd might call bash or R functions to do the work.
Scatter plot matrix seems right on the money
readxl is a package in R that might be useful if there are a lot of excel files or lots of sheets in an excel file.